padding oracle attack
a bit tricky.

once we solve for the last byte of a block, we can move backwards
through the block and get all the other bytes. if we know the IV,
we can solve the first block as well.

this code relies heavily on lists and definitely be made faster using
Vector.