C20: solve using repeating-XOR style statistical approach
the approach taken here is like the repeating-XOR of C3. we truncate the ciphertexts to
the length of the shortest, then transpose them, so that the first list contains the first
byte of each ciphertext, the second list contains the seconds, and so on. we guess a key and
use an english character frequency table to score guesses. we pick the best and return the
guessed key.