C21: mersenne twister
this implementation uses the neat (and new to me) ST.Lazy monad to
create an infinite stream of random Word32s. In practice you wouldn't
want this exact behavior since you wouldn't be able to support multiple
types, but it's pretty neat.

by running a loop forever and lazily emitting values we can return a
pure list of Word32s which the user can draw from. meanwhile we hold
a MVector of Word32s and iterate across it, looping around when we go
off the end. reads are modded in this way as well, so that we generate
the sequence in place.

test vectors are generated from the 'mersenne-twister' NPM package.